+++ src/Streamly/FileSystem/FD.hs
@@ -151,6 +151,7 @@ import qualified Streamly.FileSystem.FDIO as RawIO hiding (write)
 -- import Streamly.String (encodeUtf8, decodeUtf8, foldLines)
 
 import qualified Streamly.Mem.Array as A
+import qualified Streamly.Mem.Array.Stream as AS
 import qualified Streamly.Prelude as S
 import qualified Streamly.Streams.StreamD.Type as D
 
@@ -326,7 +327,7 @@ readArrays = readArraysOfUpto defaultChunkSize
 --
 {-# INLINE readByChunksUpto #-}
 readByChunksUpto :: (IsStream t, MonadIO m) => Int -> Handle -> t m Word8
-readByChunksUpto chunkSize h = A.flattenArrays $ readArraysOfUpto chunkSize h
+readByChunksUpto chunkSize h = AS.flatten $ readArraysOfUpto chunkSize h
 
 -- TODO
 -- read :: (IsStream t, MonadIO m, Storable a) => Handle -> t m a
@@ -338,7 +339,7 @@ readByChunksUpto chunkSize h = A.flattenArrays $ readArraysOfUpto chunkSize h
 -- @since 0.7.0
 {-# INLINE read #-}
 read :: (IsStream t, MonadIO m) => Handle -> t m Word8
-read = A.flattenArrays . readArrays
+read = AS.flatten . readArrays
 
 -------------------------------------------------------------------------------
 -- Writing
@@ -360,7 +361,7 @@ writeArrays h m = S.mapM_ (liftIO . writeArray h) m
 {-# INLINE writeArraysPackedUpto #-}
 writeArraysPackedUpto :: (MonadIO m, Storable a)
     => Int -> Handle -> SerialT m (Array a) -> m ()
-writeArraysPackedUpto n h xs = writeArrays h $ A.packArraysChunksOf n xs
+writeArraysPackedUpto n h xs = writeArrays h $ AS.compact n xs
 
 #if !defined(mingw32_HOST_OS)
 -- XXX this is incomplete
@@ -400,7 +401,7 @@ _writevArraysPackedUpto n h xs =
 -- @since 0.7.0
 {-# INLINE writeByChunksOf #-}
 writeByChunksOf :: MonadIO m => Int -> Handle -> SerialT m Word8 -> m ()
-writeByChunksOf n h m = writeArrays h $ A.arraysOf n m
+writeByChunksOf n h m = writeArrays h $ AS.arraysOf n m
 
 -- > write = 'writeByChunks' A.defaultChunkSize
 --