+++ src/Streamly/Mem/Array.hs
@@ -122,22 +122,11 @@ module Streamly.Mem.Array
     , readSliceRev
     -}
 
-    , writeIndex
+    -- , writeIndex
     {-
     , writeSlice
     , writeSliceRev
     -}
-
-    -- * Streams of arrays
-    , arraysOf
-
-    -- Streams of arrays
-    , flattenArrays
-    -- , flattenArraysRev
-    , packArrays
-    , packArraysChunksOf
-    , unlinesArraysBy
-    , splitArraysOn
 
     -- * Immutable Transformations
     , transformWith
@@ -151,10 +140,9 @@ where
 import Control.Monad (when)
 import Control.Monad.IO.Class (MonadIO(..))
 -- import Data.Functor.Identity (Identity)
-import Data.Word (Word8)
 import Foreign.ForeignPtr (withForeignPtr)
 import Foreign.ForeignPtr.Unsafe (unsafeForeignPtrToPtr)
-import Foreign.Ptr (minusPtr, plusPtr, castPtr)
+import Foreign.Ptr (minusPtr, plusPtr)
 import Foreign.Storable (Storable(..))
 import Prelude hiding (length, null, last, map, (!!), read)
 
@@ -165,9 +153,7 @@ import Streamly.Fold.Types (Fold(..))
 
 import qualified Streamly.Fold as FL
 import qualified Streamly.Mem.Array.Types as A
-import qualified Streamly.Prelude as S
 import qualified Streamly.Streams.StreamD as D
-import qualified Streamly.Streams.Prelude as P
 
 -------------------------------------------------------------------------------
 -- Construction
@@ -351,9 +337,9 @@ readSliceRev arr i len = undefined
 -- Performs in-place mutation of the array.
 --
 -- @since 0.7.0
-{-# INLINE writeIndex #-}
-writeIndex :: (MonadIO m, Storable a) => Array a -> Int -> a -> m ()
-writeIndex arr i a = do
+{-# INLINE _writeIndex #-}
+_writeIndex :: (MonadIO m, Storable a) => Array a -> Int -> a -> m ()
+_writeIndex arr i a = do
     let maxIndex = length arr - 1
     if i < 0
     then error "writeIndex: negative array index"
@@ -422,12 +408,6 @@ _toArraysOf n = FL.lchunksOf n (A.toArrayN n) FL.toStream
 -- XXX check if GHC's memory allocator is efficient enough. We can try the C
 -- malloc to compare against.
 
-{-# INLINE bytesToCount #-}
-bytesToCount :: Storable a => a -> Int -> Int
-bytesToCount x n =
-    let elemSize = sizeOf x
-    in n + elemSize - 1 `div` elemSize
-
 {-# INLINE_NORMAL toArrayMinChunk #-}
 toArrayMinChunk :: forall m a. (MonadIO m, Storable a)
     => Int -> Fold m a (Array a)
@@ -459,124 +439,8 @@ toArrayMinChunk elemCount = Fold step initial extract
 -- @since 0.7.0
 {-# INLINE toArray #-}
 toArray :: forall m a. (MonadIO m, Storable a) => Fold m a (Array a)
-toArray = toArrayMinChunk (bytesToCount (undefined :: a) (A.mkChunkSize 1024))
-
--- | Convert a stream of arrays into a stream of their elements.
---
--- Same as the following but more efficient:
---
--- > flattenArrays = S.concatMap A.read
---
--- @since 0.7.0
-{-# INLINE flattenArrays #-}
-flattenArrays :: (IsStream t, MonadIO m, Storable a) => t m (Array a) -> t m a
-flattenArrays m = D.fromStreamD $ A.flattenArrays (D.toStreamD m)
-
--- XXX should we have a reverseArrays API to reverse the stream of arrays
--- instead?
---
--- | Convert a stream of arrays into a stream of their elements reversing the
--- contents of each array before flattening.
---
--- @since 0.7.0
-{-# INLINE _flattenArraysRev #-}
-_flattenArraysRev :: (IsStream t, MonadIO m, Storable a) => t m (Array a) -> t m a
-_flattenArraysRev m = D.fromStreamD $ A.flattenArraysRev (D.toStreamD m)
-
--- XXX use an Array instead as separator? Or use a separate unlinesArraysBySeq
--- API for that?
---
--- | Flatten a stream of arrays appending the given element after each
--- array.
---
--- @since 0.7.0
-{-# INLINE unlinesArraysBy #-}
-unlinesArraysBy :: (MonadIO m, IsStream t, Storable a)
-    => a -> t m (Array a) -> t m a
-unlinesArraysBy x = D.fromStreamD . A.unlines x . D.toStreamD
-
--- | Split a stream of arrays on a given separator byte, dropping the separator
--- and coalescing all the arrays between two separators into a single array.
---
--- @since 0.7.0
-{-# INLINE splitArraysOn #-}
-splitArraysOn
-    :: (IsStream t, MonadIO m)
-    => Word8
-    -> t m (Array Word8)
-    -> t m (Array Word8)
-splitArraysOn byte s = D.fromStreamD $ A.splitOn byte $ D.toStreamD s
-
--- | Coalesce adajcent arrays in incoming stream to form bigger arrays of a
--- maximum specified size.
---
--- @since 0.7.0
-{-# INLINE packArraysChunksOf #-}
-packArraysChunksOf :: (MonadIO m, Storable a)
-    => Int -> SerialT m (Array a) -> SerialT m (Array a)
-packArraysChunksOf n xs =
-    D.fromStreamD $ A.packArraysChunksOf n (D.toStreamD xs)
-
--- | Groups the elements in an input stream into arrays of given size.
---
--- Same as the following but more efficient:
---
--- > arraysOf n = FL.groupsOf n (FL.toArrayN n)
---
--- @since 0.7.0
-{-# INLINE arraysOf #-}
-arraysOf :: (IsStream t, MonadIO m, Storable a)
-    => Int -> t m a -> t m (Array a)
-arraysOf n str =
-    D.fromStreamD $ A.fromStreamDArraysOf n (D.toStreamD str)
-
--- XXX Both of these implementations of splicing seem to perform equally well.
--- We need to perform benchmarks over a range of sizes though.
-
--- CAUTION! length must more than equal to lengths of all the arrays in the
--- stream.
-{-# INLINE spliceArraysLenUnsafe #-}
-spliceArraysLenUnsafe :: (MonadIO m, Storable a)
-    => Int -> SerialT m (Array a) -> m (Array a)
-spliceArraysLenUnsafe len buffered = do
-    arr <- liftIO $ A.newArray len
-    end <- S.foldlM' writeArr (aEnd arr) buffered
-    return $ arr {aEnd = end}
-
-    where
-
-    writeArr dst Array{..} =
-        liftIO $ withForeignPtr aStart $ \src -> do
-                        let count = aEnd `minusPtr` src
-                        A.memcpy (castPtr dst) (castPtr src) count
-                        return $ dst `plusPtr` count
-
-{-# INLINE _spliceArraysBuffered #-}
-_spliceArraysBuffered :: (MonadIO m, Storable a)
-    => SerialT m (Array a) -> m (Array a)
-_spliceArraysBuffered s = do
-    buffered <- P.foldr S.cons S.nil s
-    len <- S.sum (S.map length buffered)
-    spliceArraysLenUnsafe len s
-
-{-# INLINE spliceArraysRealloced #-}
-spliceArraysRealloced :: forall m a. (MonadIO m, Storable a)
-    => SerialT m (Array a) -> m (Array a)
-spliceArraysRealloced s = do
-    idst <- liftIO $ A.newArray (bytesToCount (undefined :: a)
-                                (A.mkChunkSizeKB 4))
-
-    arr <- S.foldlM' A.spliceWithDoubling idst s
-    liftIO $ A.shrinkToFit arr
-
--- | Given a stream of arrays, splice them all together to generate a single
--- array. The stream must be /finite/.
---
--- @since 0.7.0
-{-# INLINE packArrays #-}
-packArrays :: (MonadIO m, Storable a) => SerialT m (Array a) -> m (Array a)
-packArrays = spliceArraysRealloced
--- spliceArrays = _spliceArraysBuffered
+toArray =
+    toArrayMinChunk (A.bytesToCount (undefined :: a) (A.mkChunkSize 1024))
 
 -- | Create an 'Array' from a stream. This is useful when we want to create a
 -- single array from a stream of unknown size. 'writeN' is at least twice